+++ MSVC/suanPan/suanPan/suanPan.vcxproj
@@ -558,12 +558,12 @@
     <ClInclude Include="..\..\..\Domain\MetaMat\DenseMat.hpp" />
     <ClInclude Include="..\..\..\Domain\MetaMat\FullMat.hpp" />
     <ClInclude Include="..\..\..\Domain\MetaMat\FullMatCUDA.hpp" />
+    <ClInclude Include="..\..\..\Domain\MetaMat\ILU.hpp" />
     <ClInclude Include="..\..\..\Domain\MetaMat\IterativeSolver.hpp" />
+    <ClInclude Include="..\..\..\Domain\MetaMat\Jacobi.hpp" />
     <ClInclude Include="..\..\..\Domain\MetaMat\MetaMat.hpp" />
     <ClInclude Include="..\..\..\Domain\MetaMat\operator_times.hpp" />
-    <ClInclude Include="..\..\..\Domain\MetaMat\Preconditioner\ILU.hpp" />
-    <ClInclude Include="..\..\..\Domain\MetaMat\Preconditioner\Jacobi.hpp" />
-    <ClInclude Include="..\..\..\Domain\MetaMat\Preconditioner\Preconditioner.hpp" />
+    <ClInclude Include="..\..\..\Domain\MetaMat\Preconditioner.hpp" />
     <ClInclude Include="..\..\..\Domain\MetaMat\SolverSetting.hpp" />
     <ClInclude Include="..\..\..\Domain\MetaMat\SparseMat.hpp" />
     <ClInclude Include="..\..\..\Domain\MetaMat\SparseMatCUDA.hpp" />
