+++ MSVC/suanPan/suanPan/suanPan.vcxproj.filters
@@ -202,9 +202,6 @@
     <Filter Include="Domain\Group">
       <UniqueIdentifier>{ea4086e5-ad87-4755-905a-5f572935c34d}</UniqueIdentifier>
     </Filter>
-    <Filter Include="Domain\MetaMat\Preconditioner">
-      <UniqueIdentifier>{a0b16f39-4f62-419f-9dd1-c93b961a385d}</UniqueIdentifier>
-    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\..\Constraint\BC\GroupMultiplierBC.cpp">
@@ -2650,14 +2647,14 @@
     <ClInclude Include="..\..\..\Domain\MetaMat\IterativeSolver.hpp">
       <Filter>Domain\MetaMat</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\Domain\MetaMat\Preconditioner\ILU.hpp">
-      <Filter>Domain\MetaMat\Preconditioner</Filter>
+    <ClInclude Include="..\..\..\Domain\MetaMat\Preconditioner.hpp">
+      <Filter>Domain\MetaMat</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\Domain\MetaMat\Preconditioner\Jacobi.hpp">
-      <Filter>Domain\MetaMat\Preconditioner</Filter>
+    <ClInclude Include="..\..\..\Domain\MetaMat\ILU.hpp">
+      <Filter>Domain\MetaMat</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\..\Domain\MetaMat\Preconditioner\Preconditioner.hpp">
-      <Filter>Domain\MetaMat\Preconditioner</Filter>
+    <ClInclude Include="..\..\..\Domain\MetaMat\Jacobi.hpp">
+      <Filter>Domain\MetaMat</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
